refactor: Divide endpoints into separate files
Split endpoints into separate files, grouped by functionality.
